// frontend/TransportesOrellanaSpa.Web/src/app/core/models/actualizar-viaje.ts
export interface ActualizarViaje {
  numeroGuiaDespacho: string;
  fecha: string;

  clienteId: number;
  camionId: number;
  conductorId: number;
  remolqueId: number;

  origen: string;
  destino: string;
  comunaOrigen: string;
  comunaDestino: string;
  tipoCarga: string;

  kilometros: number | null;

  litrosCombustible: number;
  costoCombustible: number;
  tarifa: number;

  observaciones: string;

  estado: string;
}